seperate test for pi circuit
