+++ zkevm-circuits/src/pi_circuit/param.rs
@@ -1,0 +1,5 @@
+/// Fixed by the spec
+pub(super) const BLOCK_LEN: usize = 7 + 256;
+pub(super) const EXTRA_LEN: usize = 2;
+pub(super) const ZERO_BYTE_GAS_COST: u64 = 4;
+pub(super) const NONZERO_BYTE_GAS_COST: u64 = 16;
